+++ LootCollector/Modules/Settings.lua
@@ -1510,7 +1510,7 @@ local function buildOptions()
                         name = "• GitHub Sponsors",
                         order = 43,
                         func = function()
-                            OpenCopyPopup("GitHub Sponsors", "https://github.com/sponsors/mmobrain")
+                            OpenCopyPopup("GitHub Sponsors", "https://github.com/sponsors/gerob")
                         end,
                     },
                     ingame_mail_desc = {